Commented out lines 4,5,6
Commented out these lines because I think they were causing my site to point to a ERROR 500 and where not allowing me to access my wp-admin access login.  What's interesting is that I commented out the lines in y htaccess file and my site still redirects to the SSL version of the website.